finalized model Port
uuid is not working. the id is, for some reason, not being generated in
the db.
created a composite pk out of switchId and name instead.
added @JoinColumn so that the column name is not "switch_id_id".
added a host_mac_id column with a oneToOne realtion in oder so see which
host is connected to which port.

added sample data for port.